sx1276: fix variable size

+++ drivers/sx1276/sx1276.c
@@ -1114,7 +1114,7 @@ void sx1276_read_fifo(sx1276_t *dev, uint8_t *buffer, uint8_t size)
 }
 
 sx1276_modem_status_t sx1276_get_modem_status(sx1276_t *dev) {
-    uint32_t status = sx1276_reg_read(dev, REG_LR_MODEMSTAT);
+    uint8_t status = sx1276_reg_read(dev, REG_LR_MODEMSTAT);
     if (status & (1 << 4)) {
         return SX1276_MODEM_CLEAR;
     }
